Throw NullPointerException
If null is a valid return value from the server, then perhaps the
execute() method should be allowed to sometimes return a null. Perhaps
queryResultToLong() and queryResultToDouble() - which are only used by
the easy count/max/min/etc. methods - should throw NullPointerException
with a custom message if the result is null?