Initial go at an implementation of Funnel, unit tests still to come.
Includes some minor refactoring in KeenQueryClient to make execute more generic and move URL generation to its own class which is used by the KeenRequest implementation to get a URL specific to the type of request.

A minor change was made to the Query.Builder interface. The list of filters is now of type List<Filter> to make use of the new Filter object and reduce code duplication.